+++ middleman-core/lib/middleman-core/sitemap/store.rb
@@ -137,7 +137,7 @@ module Middleman
       # @param [String] source_file
       # @return [Hash]
       def metadata_for_file(source_file)
-        blank_metadata = { :options => {}, :locals => {}, :page => {} }
+        blank_metadata = { :options => {}, :locals => {} }
 
         provides_metadata.inject(blank_metadata) do |result, (callback, matcher)|
           next result if matcher && !source_file.match(matcher)
@@ -165,7 +165,7 @@ module Middleman
       def metadata_for_path(request_path)
         return @_cached_metadata[request_path] if @_cached_metadata[request_path]
 
-        blank_metadata = { :options => {}, :locals => {}, :page => {} }
+        blank_metadata = { :options => {}, :locals => {} }
 
         @_cached_metadata[request_path] = provides_metadata_for_path.inject(blank_metadata) do |result, (callback, matcher)|
           case matcher